requirements.txt
streamlit==1.32.0
pandas==2.2.1
spotipy==2.23.0
python-dotenv==1.0.1
requests==2.31.0 